perf: break component gallery and optimize logic loading image
