+++ apps/mobile/src/app/screens/home/homedrawer/components/AttachmentPicker/Gallery/components/GalleryItem.tsx
@@ -1,0 +1,82 @@
+import { CameraIcon, CheckIcon, PlayIcon } from '@mezon/mobile-components';
+import { Colors, size } from '@mezon/mobile-ui';
+import { PreSendAttachment } from '@mezon/utils';
+import { PhotoIdentifier } from '@react-native-camera-roll/camera-roll';
+import { useState } from 'react';
+import { ActivityIndicator, TouchableOpacity, View } from 'react-native';
+import FastImage from 'react-native-fast-image';
+import { style } from './styles';
+
+interface GalleryItemProps {
+	item: any;
+	themeValue: any;
+	isDisableSelectAttachment: boolean;
+	attachmentFilteredByChannelId: PreSendAttachment;
+	onOpenCamera: () => void;
+	handleGalleryPress: (file: PhotoIdentifier) => Promise<void>;
+	handleRemove: (filename: string) => void;
+}
+
+const GalleryItem = ({
+	item,
+	themeValue,
+	isDisableSelectAttachment,
+	attachmentFilteredByChannelId,
+	onOpenCamera,
+	handleGalleryPress,
+	handleRemove
+}: GalleryItemProps) => {
+	const styles = style(themeValue);
+
+	if (item?.isUseCamera) {
+		return (
+			<TouchableOpacity style={[styles.cameraPicker]} onPress={onOpenCamera}>
+				<CameraIcon color={themeValue.text} width={size.s_24} height={size.s_24} />
+			</TouchableOpacity>
+		);
+	}
+
+	const fileName = item?.node?.image?.filename;
+	const isVideo = item?.node?.type?.startsWith?.('video');
+	const isSelected = attachmentFilteredByChannelId?.files.some((file) => file.filename === fileName);
+	const disabled = isDisableSelectAttachment && !isSelected;
+	const [isLoadingImage, setIsLoadingImage] = useState(true);
+
+	return (
+		<TouchableOpacity
+			style={[styles.itemGallery, disabled && styles.disable]}
+			onPress={() => {
+				if (isSelected) {
+					handleRemove(fileName);
+				} else {
+					handleGalleryPress(item);
+				}
+			}}
+			disabled={disabled}
+		>
+			<FastImage
+				source={{ uri: item.node.image.uri + '?thumbnail=true&quality=low', cache: FastImage.cacheControl.immutable }}
+				style={styles.imageGallery}
+				onLoadEnd={() => setIsLoadingImage(false)}
+			/>
+			{isLoadingImage && (
+				<View style={styles.loadingContainer}>
+					<ActivityIndicator size="small" color={themeValue.text} />
+				</View>
+			)}
+			{isVideo && (
+				<View style={styles.videoOverlay}>
+					<PlayIcon width={size.s_20} height={size.s_20} />
+				</View>
+			)}
+			{isSelected && (
+				<View style={styles.iconSelected}>
+					<CheckIcon color={Colors.bgViolet} />
+				</View>
+			)}
+			{isSelected && <View style={styles.selectedOverlay} />}
+		</TouchableOpacity>
+	);
+};
+
+export default GalleryItem;
