+++ apps/mobile/src/app/screens/home/homedrawer/components/AttachmentPicker/Gallery/index.tsx
@@ -1,23 +1,12 @@
 import { useReference } from '@mezon/core';
-import { ActionEmitEvent, CameraIcon, CheckIcon, PlayIcon } from '@mezon/mobile-components';
-import { Colors, size, useTheme } from '@mezon/mobile-ui';
+import { ActionEmitEvent } from '@mezon/mobile-components';
+import { useTheme } from '@mezon/mobile-ui';
 import { appActions, useAppDispatch } from '@mezon/store-mobile';
 import { MAX_FILE_SIZE } from '@mezon/utils';
 import { CameraRoll, PhotoIdentifier, iosRefreshGallerySelection, iosRequestReadWriteGalleryPermission } from '@react-native-camera-roll/camera-roll';
 import { iosReadGalleryPermission } from '@react-native-camera-roll/camera-roll/src/CameraRollIOSPermission';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
-import {
-	ActivityIndicator,
-	Alert,
-	DeviceEventEmitter,
-	Dimensions,
-	Linking,
-	PermissionsAndroid,
-	Platform,
-	TouchableOpacity,
-	View
-} from 'react-native';
-import FastImage from 'react-native-fast-image';
+import { ActivityIndicator, Alert, DeviceEventEmitter, Dimensions, Linking, PermissionsAndroid, Platform, View } from 'react-native';
 import RNFS from 'react-native-fs';
 import { FlatList } from 'react-native-gesture-handler';
 import * as ImagePicker from 'react-native-image-picker';
@@ -25,7 +14,7 @@ import { CameraOptions } from 'react-native-image-picker';
 import Toast from 'react-native-toast-message';
 import { Camera } from 'react-native-vision-camera';
 import { IFile } from '../../../../../../componentUI/MezonImagePicker';
-import { style } from './styles';
+import GalleryItem from './components/GalleryItem';
 
 export const { height } = Dimensions.get('window');
 interface IProps {
@@ -35,12 +24,10 @@ interface IProps {
 
 const Gallery = ({ onPickGallery, currentChannelId }: IProps) => {
 	const { themeValue } = useTheme();
-	const styles = style(themeValue);
 	const [photos, setPhotos] = useState<PhotoIdentifier[]>([]);
 	const [currentAlbums, setCurrentAlbums] = useState<string>('All');
 	const [pageInfo, setPageInfo] = useState(null);
 	const [isLoadingMore, setIsLoadingMore] = useState(false);
-	const [loadingImages, setLoadingImages] = useState<Record<string, boolean>>({});
 	const dispatch = useAppDispatch();
 	const timerRef = useRef<any>();
 	const { removeAttachmentByIndex, attachmentFilteredByChannelId } = useReference(currentChannelId);
@@ -187,20 +174,12 @@ const Gallery = ({ onPickGallery, currentChannelId }: IProps) => {
 		setIsLoadingMore(true);
 		try {
 			const res = await CameraRoll.getPhotos({
-				first: 30,
+				first: 32,
 				assetType: album === 'All Videos' ? 'Videos' : 'All',
 				...(!!pageInfo && !!after && { after: after }),
 				include: ['filename', 'fileSize', 'fileExtension', 'imageSize', 'orientation'],
 				groupTypes: album === 'All' ? 'All' : 'Album',
 				groupName: album === 'All' || album === 'All Videos' ? null : album
-			});
-
-			setLoadingImages((prevState) => {
-				const newState = { ...prevState };
-				res.edges.forEach((photo) => {
-					newState[photo.node.id] = true;
-				});
-				return newState;
 			});
 
 			setPhotos(after ? [...photos, ...res.edges] : res.edges);
@@ -223,56 +202,16 @@ const Gallery = ({ onPickGallery, currentChannelId }: IProps) => {
 	}, []);
 
 	const renderItem = ({ item }) => {
-		if (item?.isUseCamera) {
-			return (
-				<TouchableOpacity style={[styles.cameraPicker]} onPress={onOpenCamera}>
-					<CameraIcon color={themeValue.text} width={size.s_24} height={size.s_24} />
-				</TouchableOpacity>
-			);
-		}
-		const fileName = item?.node?.image?.filename;
-		const imageId = item?.node?.id;
-		const isVideo = item?.node?.type?.startsWith?.('video');
-		const isSelected = attachmentFilteredByChannelId?.files.some((file) => file.filename === fileName);
-		const disabled = isDisableSelectAttachment && !isSelected;
-		const isLoadingImage = loadingImages[imageId];
-
 		return (
-			<TouchableOpacity
-				style={[styles.itemGallery, disabled && styles.disable]}
-				onPress={() => {
-					if (isSelected) {
-						handleRemove(fileName);
-					} else {
-						handleGalleryPress(item);
-					}
-				}}
-				disabled={disabled}
-			>
-				<FastImage
-					source={{ uri: item.node.image.uri + '?thumbnail=true&quality=low', cache: FastImage.cacheControl.immutable }}
-					style={styles.imageGallery}
-					onLoadEnd={() => {
-						setLoadingImages((prev) => ({ ...prev, [imageId]: false }));
-					}}
-				/>
-				{isLoadingImage && (
-					<View style={styles.loadingContainer}>
-						<ActivityIndicator size="small" color={themeValue.text} />
-					</View>
-				)}
-				{isVideo && (
-					<View style={styles.videoOverlay}>
-						<PlayIcon width={size.s_20} height={size.s_20} />
-					</View>
-				)}
-				{isSelected && (
-					<View style={styles.iconSelected}>
-						<CheckIcon color={Colors.bgViolet} />
-					</View>
-				)}
-				{isSelected && <View style={styles.selectedOverlay} />}
-			</TouchableOpacity>
+			<GalleryItem
+				item={item}
+				themeValue={themeValue}
+				isDisableSelectAttachment={isDisableSelectAttachment}
+				attachmentFilteredByChannelId={attachmentFilteredByChannelId}
+				onOpenCamera={onOpenCamera}
+				handleGalleryPress={handleGalleryPress}
+				handleRemove={handleRemove}
+			/>
 		);
 	};
